% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitterwidget.R
\name{twitterwidget}
\alias{twitterwidget}
\title{Create a widget representing a twitter status}
\usage{
twitterwidget(twid, pars = NULL, width = NULL, height = NULL,
  elementId = NULL)
}
\description{
The method create a widget able to embed a twitter status within a
Rmd page. The main parameter is the status ID, other parameters will
be passed to the factory method, according the official documentation:
https://developer.twitter.com/en/docs/twitter-for-websites/embedded-tweets/guides/embedded-tweet-parameter-reference
}
