% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitterwidget.R
\name{twitterwidget}
\alias{twitterwidget}
\title{Create a widget representing a twitter status}
\usage{
twitterwidget(twid, pars = NULL, width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{twid}{ID of the twitter status, it should be passed as a string}

\item{pars}{list() containing the optional parameters passed to request
the rendering of the embedded tweets.}

\item{width}{width of the widget}

\item{height}{height of the widget}

\item{elementId}{ID of the HTML element}
}
\description{
The method create a widget able to embed a twitter status within a
Rmd page. The main parameter is the status ID, other parameters will
be passed to the factory method, according the official documentation:
https://developer.twitter.com/en/docs/twitter-for-websites/embedded-tweets/guides/embedded-tweet-parameter-reference
}
